chore: refactor TeamContexts to separate file
